--- frontend/src/components/Layout.module.css
@@ -1,10 +1,0 @@
-.layout {
-  display: flex;
-  min-height: 100vh;
-  background: var(--bg-900);
-}
-
-.main {
-  flex: 1;
-  padding: 20px;
-}
