+++ frontend/src/components/menu/Button.module.css
@@ -1,0 +1,33 @@
+.btn {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  padding: 10px 16px;
+  border-radius: 10px;
+  border: none;
+  cursor: pointer;
+  font-weight: 600;
+}
+
+/* Variantes */
+.primary {
+  background: var(--accent-100);
+  color: var(--bg-900);
+}
+
+.secondary {
+  background: var(--bg-900);
+  color: white;
+}
+
+.ghost {
+  background: transparent;
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  color: var(--accent-200);
+}
+
+.outline {
+  background: transparent;
+  border: 2px solid var(--accent-300);
+  color: var(--accent-300);
+}
