+++ frontend/src/components/menu/Layout.module.css
@@ -1,0 +1,10 @@
+.layout {
+  display: flex;
+  min-height: 100vh;
+  background: var(--white);
+}
+
+.main {
+  flex: 1;
+  padding: 20px;
+}
