+++ frontend/src/components/menu/Sidebar.module.css
@@ -1,0 +1,49 @@
+.sidebar {
+  width: 220px;
+  background: var(--accent-400);
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.header {
+  margin-bottom: 20px;
+}
+
+.title {
+  margin: 0;
+  font-size: 18px;
+  font-weight: bold;
+  color: black;
+}
+
+.subtitle {
+  font-size: 12px;
+  color: black;
+}
+
+.nav {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.link {
+  text-decoration: none;
+  color: #94a5a7;
+  font-weight: 500;
+  padding: 6px 10px;
+  border-radius: 6px;
+  transition: background 0.2s;
+}
+
+.link:hover {
+  background: #d5dfe1;
+
+}
+
+.active {
+  background: var(--accent-100);
+  color: var(--bg-900);
+}
